Refactor User (actor) to UserActor to avoid confusing with User (response)
